+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.14.6] - 2026-02-05
+
+### Fixed
+- Yale cloud locks now filtered at discovery time (no SSE events, minimal API data)
+
+## [0.14.5] - 2026-02-05
+
+### Added
+- New `ajax.get_smart_locks` diagnostic service to fetch smart lock data from API and SSE/SQS-discovered locks
+
 ## [0.14.4] - 2026-02-04
 
 ### Added